requirements.txt
google-genai>=0.6.0
python-dotenv
requests
